Refactor common code into helper file
